Target_supported and mcu create belong to target file
