+++ project_generator/targets.py
@@ -19,6 +19,7 @@ from os.path import join, normpath, splitext, isfile, exists
 from os import listdir, mkdir
 
 from .settings import ProjectSettings
+from .tools_supported import ToolsSupported
 
 class Targets:
 
@@ -26,7 +27,7 @@ class Targets:
         'mcu' : {
             'vendor' : ['Manually add vendor (st, freescale, etc) instead of this text'],
             'name' : [''],
-            'core' : ['Manually add core (cortex-mX)instead of this text'],
+            'core' : ['Manually add core (cortex-mX) instead of this text'],
         },
     }
 
@@ -113,3 +114,32 @@ class Targets:
                     cmd = ('git', 'pull', '--rebase', '--quiet', 'origin', 'master')
                     subprocess.call(cmd, cwd=settings.paths['definitions'])
 
+# These 2 methods are related to Targets. One is to help with creating new
+# targets - mcu_create, the target_supported is to check if target is supported
+
+# This method checks if target is supported by default (nothing more needed)
+# or requires additional target definitions
+def target_supported(exporter, target, tool, env_settings):
+    if exporter not in ToolsSupported().get_supported():
+        logging.debug("Target does not support specified tool: %s" % tool)
+        return None
+    else:
+        supported = exporter.is_supported_by_default(target)
+        # target requires further definitions for exporter
+        if not supported:
+            Target = Targets(env_settings.get_env_settings('definitions'))
+            supported = Target.is_supported(target, tool)
+        return supported
+
+# This helps to create a new target. As target consists of mcu, this function
+# parses the provided proj_file and creates a valid yaml file, which can be pushed
+# to pgen definitions.
+def mcu_create(ToolParser, mcu_name, proj_file, tool):
+    data = ToolParser(None, None).get_mcu_definition(proj_file)
+    data['mcu']['name'] = [mcu_name]
+    # we got target, now damp it to root using target.yaml file
+    # we can make it better, and ask for definitions repo clone, and add it
+    # there, at least to MCU folder
+    with open(os.path.join(os.getcwd(), mcu_name + '.yaml'), 'wt') as f:
+        f.write(yaml.safe_dump(data, default_flow_style=False, width=200))
+    return 0
