+++ project_generator/tools_supported.py
@@ -12,11 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-import logging
-import subprocess
-import yaml
-
 from .targets import Targets
 from .tools.iar import IAREmbeddedWorkbench
 from .tools.uvision import Uvision
@@ -70,23 +65,3 @@ class ToolsSupported:
     def get_supported(self):
         return self.TOOLS_DICT.keys()
 
-def target_supported(exporter, target, tool, env_settings):
-    if exporter not in ToolsSupported().get_supported():
-        raise RuntimeError("Target does not support specified tool: %s" % tool)
-    else:
-        supported = exporter.is_supported_by_default(target)
-        # target requires further definitions for exporter
-        if not supported:
-            Target = Targets(env_settings.get_env_settings('definitions'))
-            supported = Target.is_supported(target, tool)
-        return supported
-
-def mcu_create(ToolParser, mcu_name, proj_file, tool):
-    data = ToolParser(None, None).get_mcu_definition(proj_file)
-    data['mcu']['name'] = [mcu_name]
-    # we got target, now damp it to root using target.yaml file
-    # we can make it better, and ask for definitions repo clone, and add it
-    # there, at least to MCU folder
-    with open(os.path.join(os.getcwd(), mcu_name + '.yaml'), 'wt') as f:
-        f.write(yaml.safe_dump(data, default_flow_style=False, width=200))
-    return 0
